+++ modules/sensors/veml7700.c
@@ -2,7 +2,7 @@
 #include "veml7700.h"
 #include "driver/i2c.h"
 #include "esp_log.h"
-#include <math.h> // Needed for the correction formula (powf)
+#include <math.h>
 
 static const char *TAG = "VEML7700";
 
@@ -33,7 +33,6 @@ static const char *TAG = "VEML7700";
 #define COEF_C 8.1488e-5
 #define COEF_D 1.0023
 
-// --- Helper: Write 16-bit word ---
 static esp_err_t write_reg(uint8_t reg, uint16_t val)
 {
     uint8_t data[3];
@@ -43,7 +42,6 @@ static esp_err_t write_reg(uint8_t reg, uint16_t val)
     return i2c_master_write_to_device(I2C_PORT, VEML7700_ADDR, data, 3, pdMS_TO_TICKS(100));
 }
 
-// --- Helper: Read 16-bit word ---
 static esp_err_t read_reg(uint8_t reg, uint16_t *val)
 {
     uint8_t raw[2];
@@ -55,7 +53,6 @@ static esp_err_t read_reg(uint8_t reg, uint16_t *val)
     return ret;
 }
 
-// --- Public: Init ---
 esp_err_t veml7700_init(int sda_pin, int scl_pin)
 {
     // 1. Setup I2C Config
@@ -67,17 +64,6 @@ esp_err_t veml7700_init(int sda_pin, int scl_pin)
         .scl_pullup_en = GPIO_PULLUP_ENABLE,
         .master.clk_speed = I2C_FREQ_HZ,
     };
-
-    // Install driver (checks if already installed to be safe)
-    // if (i2c_driver_install(I2C_PORT, conf.mode, 0, 0, 0) != ESP_OK)
-    // {
-    //     i2c_param_config(I2C_PORT, &conf);
-    //     i2c_driver_install(I2C_PORT, conf.mode, 0, 0, 0);
-    // }
-    // else
-    // {
-    //     i2c_param_config(I2C_PORT, &conf);
-    // }
 
     // 2. Configure Sensor
     // Value: Gain 1/8 | IT 100ms | No Interrupts | Power On (Shutdown bit = 0)
@@ -92,23 +78,19 @@ esp_err_t veml7700_init(int sda_pin, int scl_pin)
         return ret;
     }
 
-    // Wait 5ms for power up (as per datasheet)
     vTaskDelay(pdMS_TO_TICKS(10));
     ESP_LOGI(TAG, "Initialized. Gain: 1/8, IT: 100ms");
     return ESP_OK;
 }
 
-// --- Public: Read Lux ---
 esp_err_t veml7700_read_lux(float *lux_val)
 {
     uint16_t raw_counts = 0;
 
-    // 1. Read raw data
     esp_err_t ret = read_reg(CMD_ALS_DATA, &raw_counts);
     if (ret != ESP_OK)
         return ret;
 
-    // 2. Convert to basic Lux
     float lux = raw_counts * LUX_RESOLUTION;
 
     // 3. Apply non-linear correction (Required by datasheet for VEML7700)
@@ -141,12 +123,10 @@ void veml7700_task(void *arg)
     while (1)
     {
         float lux = 0.0;
-
-        // This is the function you wanted
         if (veml7700_read_lux(&lux) == ESP_OK)
         {
             // ESP_LOGI(TAG, "Light Level: %.2f Lux", lux);
-            *(double *)arg = lux;
+            *(float *)arg = lux;
             vTaskDelay(pdMS_TO_TICKS(3600000));
         }
         else
@@ -157,7 +137,7 @@ void veml7700_task(void *arg)
     }
 }
 
-void veml7700_start_task(double *parameter)
+void veml7700_start_task(float *parameter)
 {
     xTaskCreate(veml7700_task, "VEML7700_Task", 4096, parameter, 10, NULL);
 }
